# docs/swagger.yaml
basePath: /
definitions:
  CreateAddressRequest:
    properties:
      external_id:
        type: string
      label:
        type: string
      lat:
        type: number
      lng:
        type: number
      name:
        type: string
    required:
      - label
      - lat
      - lng
      - name
    type: object
  UpdateContactRequest:
    properties:
      about:
        type: string
      activity_name:
        type: string
      avatar:
        type: string
      country_code:
        type: string
      email:
        type: string
      name:
        type: string
      phone_calling_code:
        type: string
      phone_number:
        type: string
      social_links:
        items:
          $ref: '#/definitions/touchly_internal_db.Link'
        type: array
      tags:
        items:
          $ref: '#/definitions/touchly_internal_db.Tag'
        type: array
      website:
        type: string
    required:
      - name
    type: object
  UploadURL:
    properties:
      url:
        type: string
    type: object
  User:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      email_verified_at:
        type: string
      id:
        type: integer
      updated_at:
        type: string
    type: object
  internal_handler.DeleteSavedContactRequest:
    properties:
      contact_id:
        example: 1
        type: integer
    type: object
  internal_handler.LoginUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  internal_handler.SendOTPRequest:
    properties:
      email:
        type: string
    type: object
  internal_handler.SetPasswordRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  internal_handler.UpdateContactVisibilityRequest:
    properties:
      visibility:
        allOf:
          - $ref: '#/definitions/touchly_internal_db.ContactVisibility'
        example: public
    type: object
  internal_handler.VerifyOTPRequest:
    properties:
      email:
        type: string
      otp:
        type: string
    type: object
  touchly_internal_db.Address:
    properties:
      contact_id:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: string
      external_id:
        type: string
      id:
        type: integer
      label:
        type: string
      location:
        $ref: '#/definitions/touchly_internal_db.Point'
      name:
        type: string
      updated_at:
        type: string
    type: object
  touchly_internal_db.Contact:
    properties:
      about:
        type: string
      activity_name:
        type: string
      address:
        $ref: '#/definitions/touchly_internal_db.Address'
      avatar:
        type: string
      country_code:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone_calling_code:
        type: string
      phone_number:
        type: string
      saves_amount:
        type: integer
      social_links:
        items:
          $ref: '#/definitions/touchly_internal_db.Link'
        type: array
      tags:
        items:
          $ref: '#/definitions/touchly_internal_db.Tag'
        type: array
      updated_at:
        type: string
      user_id:
        type: integer
      views_amount:
        type: integer
      visibility:
        $ref: '#/definitions/touchly_internal_db.ContactVisibility'
      website:
        type: string
    type: object
  touchly_internal_db.ContactListEntry:
    properties:
      about:
        type: string
      activity_name:
        type: string
      avatar:
        type: string
      id:
        type: integer
      is_saved:
        type: boolean
      name:
        type: string
      saves_amount:
        type: integer
      user_id:
        type: integer
      views_amount:
        type: integer
      visibility:
        $ref: '#/definitions/touchly_internal_db.ContactVisibility'
    type: object
  touchly_internal_db.ContactVisibility:
    enum:
      - public
      - private
      - shared_link
    type: string
    x-enum-varnames:
      - ContactVisibilityPublic
      - ContactVisibilityPrivate
      - ContactVisibilitySharedLink
  touchly_internal_db.ContactsPage:
    properties:
      contacts:
        items:
          $ref: '#/definitions/touchly_internal_db.ContactListEntry'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
    type: object
  touchly_internal_db.Link:
    properties:
      contact_id:
        type: integer
      id:
        type: integer
      label:
        type: string
      link:
        type: string
      type:
        type: string
    type: object
  touchly_internal_db.Point:
    properties:
      lat:
        type: number
      lng:
        type: number
    type: object
  touchly_internal_db.Tag:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact: { }
  description: This is a sample server ClanPlatform server.
  title: Touchly API
  version: "1.0"
paths:
  /api/contacts:
    get:
      consumes:
        - application/json
      description: get contacts
      parameters:
        - description: page number (default 1)
          in: query
          name: page
          type: integer
        - description: page size (default 20)
          in: query
          name: page_size
          type: integer
        - description: search query, search by name or activity
          in: query
          name: search
          type: string
        - collectionFormat: csv
          description: tag id
          in: query
          items:
            type: integer
          name: tag
          type: array
        - description: latitude
          in: query
          name: lat
          type: number
        - description: longitude
          in: query
          name: lng
          type: number
        - description: radius in km
          in: query
          name: radius
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/touchly_internal_db.ContactsPage'
      summary: List contacts
      tags:
        - contacts
    post:
      consumes:
        - application/json
      description: create contact
      parameters:
        - description: contact
          in: body
          name: contact
          required: true
          schema:
            $ref: '#/definitions/UpdateContactRequest'
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/touchly_internal_db.Contact'
      security:
        - JWT: [ ]
      summary: Create contact
      tags:
        - contacts
  /api/contacts/{id}:
    delete:
      consumes:
        - application/json
      description: delete contact
      parameters:
        - description: contact id
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
      security:
        - JWT: [ ]
      summary: Delete contact
      tags:
        - contacts
    get:
      consumes:
        - application/json
      description: get contact
      parameters:
        - description: contact id
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/touchly_internal_db.Contact'
      summary: Get contact
      tags:
        - contacts
    put:
      consumes:
        - application/json
      description: update contact
      parameters:
        - description: contact id
          in: path
          name: id
          required: true
          type: integer
        - description: contact
          in: body
          name: contact
          required: true
          schema:
            $ref: '#/definitions/UpdateContactRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
      security:
        - JWT: [ ]
      summary: Update contact
      tags:
        - contacts
  /api/contacts/{id}/address:
    post:
      consumes:
        - application/json
      description: create contact address
      parameters:
        - description: contact id
          in: path
          name: id
          required: true
          type: integer
        - description: address
          in: body
          name: address
          required: true
          schema:
            $ref: '#/definitions/CreateAddressRequest'
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/touchly_internal_db.Address'
      security:
        - JWT: [ ]
      summary: Create contact address
      tags:
        - contacts
  /api/contacts/{id}/save:
    delete:
      consumes:
        - application/json
      description: delete saved contact
      parameters:
        - description: user id
          in: path
          name: id
          required: true
          type: integer
        - description: contact id to delete
          in: body
          name: account
          required: true
          schema:
            $ref: '#/definitions/internal_handler.DeleteSavedContactRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
      security:
        - JWT: [ ]
      summary: Delete saved contact
      tags:
        - contacts
    post:
      consumes:
        - application/json
      description: save contact
      parameters:
        - description: contact id
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
      security:
        - JWT: [ ]
      summary: Save contact
      tags:
        - contacts
  /api/contacts/{id}/visibility:
    put:
      consumes:
        - application/json
      description: update contact visibility
      parameters:
        - description: contact id
          in: path
          name: id
          required: true
          type: integer
        - description: visibility
          in: body
          name: account
          required: true
          schema:
            $ref: '#/definitions/internal_handler.UpdateContactVisibilityRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
      security:
        - JWT: [ ]
      summary: Update contact visibility
      tags:
        - contacts
  /api/contacts/saved:
    get:
      consumes:
        - application/json
      description: get saved contacts
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/touchly_internal_db.Contact'
            type: array
      security:
        - JWT: [ ]
      summary: List contacts saved by user
      tags:
        - contacts
  /api/login:
    post:
      consumes:
        - application/json
      description: login user
      parameters:
        - description: login
          in: body
          name: login
          required: true
          schema:
            $ref: '#/definitions/internal_handler.LoginUserRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
        - users
  /api/me:
    get:
      consumes:
        - application/json
      description: get user
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
      security:
        - JWT: [ ]
      summary: Get user
      tags:
        - users
  /api/me/contacts:
    get:
      consumes:
        - application/json
      description: get my contacts
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/touchly_internal_db.ContactsPage'
      security:
        - JWT: [ ]
      summary: List my contacts
      tags:
        - contacts
  /api/otp:
    post:
      consumes:
        - application/json
      description: send OTP
      parameters:
        - description: email
          in: body
          name: email
          required: true
          schema:
            $ref: '#/definitions/internal_handler.SendOTPRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
      summary: Send OTP
      tags:
        - users
  /api/otp-verify:
    post:
      consumes:
        - application/json
      description: verify OTP
      parameters:
        - description: verify
          in: body
          name: verify
          required: true
          schema:
            $ref: '#/definitions/internal_handler.VerifyOTPRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
      summary: Verify OTP
      tags:
        - users
  /api/set-password:
    post:
      consumes:
        - application/json
      description: set password
      parameters:
        - description: password
          in: body
          name: password
          required: true
          schema:
            $ref: '#/definitions/internal_handler.SetPasswordRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
      summary: Set password
      tags:
        - users
  /api/tags:
    get:
      consumes:
        - application/json
      description: list tags
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/touchly_internal_db.Tag'
            type: array
      summary: List tags
      tags:
        - tags
    post:
      consumes:
        - application/json
      description: create tag
      parameters:
        - description: tag
          in: body
          name: tag
          required: true
          schema:
            $ref: '#/definitions/touchly_internal_db.Tag'
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/touchly_internal_db.Tag'
      security:
        - JWT: [ ]
      summary: Create tag
      tags:
        - tags
  /api/tags/{id}:
    delete:
      consumes:
        - application/json
      description: delete tag
      parameters:
        - description: tag id
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
      security:
        - JWT: [ ]
      summary: Delete tag
      tags:
        - tags
  /api/uploads/get-url:
    post:
      consumes:
        - application/json
      description: returns a presigned URL for uploading a file
      parameters:
        - description: file name
          in: query
          name: file_name
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UploadURL'
      security:
        - JWT: [ ]
      summary: Get upload URL
      tags:
        - uploads
schemes:
  - http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"